Move footer to components/navigation, remove duplicate keys on footer map menu items, create sm-tablet breakpoint for transition, use new import settings to link scss file
