+++ frontend/src/components/Navigation/Footer/_Footer.js
@@ -1,8 +1,7 @@
 import React from "react";
-import "./_Footer.scss";
-import { Button } from "../../components/Buttons/Buttons";
+import { Button } from "../../Buttons/Buttons";
 
-const Footer = ({ menu, logoDesktop, logoMobile}) => {
+const Footer = ({ menu, logoDesktop, logoMobile }) => {
   const Logo = () => {
     return (
       <div className="footer-icons-on-dark">
