+++ frontend/src/index.scss
@@ -11,6 +11,7 @@
 @use "components/Carousel/ScrollCarousel";
 @use "components/Navigation/HeaderNav";
 @use "components/Navigation/InnerCopNav";
+@use "components/Navigation/Footer/_Footer.scss";
 
 // Import page styles
